refactor: adaptive policy resource
This one was a bit more complicated in terms of behavior due to the firewall ruleset. It was required in our extraction logic to null out the list each time,  as it was a list of maps that created an opportunity for duplicate entries in our state file.

Other than that, typical refactor
 - camelCase / snake_case for applicable attributes
- renamed variables to prevent collisions throughout the provider
- changed data model to reflect team code style / strategy