chore: Peer review 2
The issue is that the Meraki OpenAPI spec states that this is an int. However, the actual return value from the API is a string.

Since we use the open API generator for the HTTP client, it is building the "port" attribute as an Int32 causing this error when we attempt to unmarshal the response:

json: cannot unmarshal string into Go struct field
        InlineResponse20065Servers.servers.port of type int32

Until the Meraki OpenAPI spec is updated with the correct type, we will use a workaround on the HTTP call, as we don't even use the openAPI type-enforced struct.

We leverage our own custom typing system to decode the *http.Response.